+++ src/main/java/com/example/Folklore/entities/FolkloreEntity.java
@@ -11,14 +11,14 @@ import lombok.NoArgsConstructor;
 public class FolkloreEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     private String title;
     private String originCountry;
     private String description;
     private String story;
     private String tags;
 
-    public FolkloreEntity(Long id, String title, String originCountry, String description, String story, String tags) {
+    public FolkloreEntity(Integer id, String title, String originCountry, String description, String story, String tags) {
         this.id = id;
         this.title = title;
         this.originCountry = originCountry;
